+++ .eslintrc.yaml
@@ -229,9 +229,6 @@ overrides:
       - "app/components/PlaybackControls/sharedHelpers.ts"
       - "app/components/PlaybackSpeedControls.tsx"
       - "app/components/Progress.tsx"
-      - "app/components/ReactChartjs/ChartJSManager.ts"
-      - "app/components/ReactChartjs/ChartJSWorker.ts"
-      - "app/components/ReactChartjs/index.tsx"
       - "app/components/Select/Option.tsx"
       - "app/components/Select/Select.tsx"
       - "app/components/SelectableTimestamp.tsx"
@@ -241,7 +238,6 @@ overrides:
       - "app/components/TextHighlight.tsx"
       - "app/components/TimeBasedChart/TimeBasedChartLegend.tsx"
       - "app/components/TimeBasedChart/TimeBasedChartTooltip.tsx"
-      - "app/components/TimeBasedChart/index.tsx"
       - "app/components/Tooltip.tsx"
       - "app/components/TooltipBase.tsx"
       - "app/components/TopicToRenderMenu.tsx"
